Update GloVe implementation
