+++ research/word_embeddings/glove/glove_kernels.cc
@@ -1,34 +1,27 @@
-#include "tensorflow/core/framework/op.h"
-#include "tensorflow/core/framework/op_kernel.h"
-#include "tensorflow/core/lib/core/stringpiece.h"
-#include "tensorflow/core/lib/strings/str_util.h"
-#include "tensorflow/core/lib/gtl/map_util.h"
+/* Copyright 2017 The TensorFlow Authors. All Rights Reserved.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+==============================================================================*/
+
+#include "../base_embedding/word_embedding_op.cc"
 
 namespace tensorflow {
 
 const int num_precalc_examples = 10000;
 
-namespace {
-
-bool ScanWord(StringPiece * input, string *word) {
-  str_util::RemoveLeadingWhitespace(input);
-  StringPiece tmp;
-
-  if (str_util::ConsumeNonWhitespace(input, &tmp)) {
-    word->assign(tmp.data(), tmp.size());
-    return true;
-  } else {
-    return false;
-  }
-}
-
-}  // end namespace
-
-
-class GloveModelOp : public OpKernel {
+class GloveModelOp : public WordEmbeddingModel {
   public:
-   explicit GloveModelOp(OpKernelConstruction* ctx)
-       : OpKernel(ctx) {
+   explicit GloveModelOp(OpKernelConstruction* ctx) : WordEmbeddingModel(ctx) {
      string filename;
      OP_REQUIRES_OK(ctx, ctx->GetAttr("filename", &filename));
      OP_REQUIRES_OK(ctx, ctx->GetAttr("window_size", &window_size_));
@@ -94,6 +87,7 @@ class GloveModelOp : public OpKernel {
     ctx->set_output(8, ccounts);
   }
 
+
   private:
    struct Example {
      int32 input;
@@ -101,124 +95,44 @@ class GloveModelOp : public OpKernel {
      float ccount;
    };
 
-   Tensor vocab_words_;
-   Tensor freq_;
    Tensor indices_;
    Tensor values_;
-   int64 corpus_size_;
-   int64 words_per_epoch_;
-   int32 window_size_;
-   int32 vocab_size_;
-   int32 min_count_;
-   int32 batch_size_;
-   std::vector<int32> corpus_;
-   std::unordered_map<uint64, float> coocurrences_;
-   std::vector<Example> precalc_examples_;
 
-   typedef std::pair<string, int32> WordFreq;
    typedef std::pair<int32, int32> CooccurIndices;
 
-   mutex mu_;
    uint64 example_pos_ GUARDED_BY(mu_) = 0;
-   int32 current_epoch_ GUARDED_BY(mu_);
-   int64 total_words_processed_ GUARDED_BY(mu_) = 0;
-   int precalc_index_ = 0 GUARDED_BY(mu_);
+   int precalc_index_ GUARDED_BY(mu_);
    std::vector<CooccurIndices> valid_indices GUARDED_BY(mu_);
+   std::vector<float> valid_values GUARDED_BY(mu_);
+   std::vector<Example> precalc_examples_ GUARDED_BY(mu_);
 
    void NextExample(int32* input, int32* label, float* ccount) EXCLUSIVE_LOCKS_REQUIRED(mu_) {
       uint64 size = static_cast<uint64>(valid_indices.size());
-      uint64 index;
-      uint64 center_word, context_word;
+      int32 center_word, context_word;
+      float value;
 
       if(example_pos_ >= size) {
         example_pos_ = 0;
         current_epoch_++;
       }
 
-      while(example_pos_++ < size) {
+      while(example_pos_ < size) {
         center_word = valid_indices[example_pos_].first;
         context_word = valid_indices[example_pos_].second;
-        index = static_cast<uint64>(center_word * vocab_size_ + context_word);
+        value = valid_values[example_pos_];
 
         *input = center_word;
         *label = context_word;
-        *ccount = coocurrences_[index];
+        *ccount = value;
         ++total_words_processed_;
+        example_pos_++;
         return;
       }
 
    }
 
-   std::unordered_map<string, int32>
-   CreateWordFrequencies(const string& data) {
-    StringPiece input = data;
-
-    string w;
-    corpus_size_ = 0;
-    std::unordered_map<string, int32> word_freq;
-    while (ScanWord(&input, &w)) {
-      ++(word_freq[w]);
-      ++corpus_size_;
-    }
-
-    return word_freq;
-   }
-
-   std::vector<WordFreq>
-   CreateVocabulary(std::unordered_map<string, int32> word_freq) {
-    std::vector<WordFreq> ordered;
-    for (const auto& p : word_freq) {
-      if (p.second >= min_count_) ordered.push_back(p);
-    }
-
-    std::sort(ordered.begin(), ordered.end(),
-              [](const WordFreq& x, const WordFreq& y) {
-                return x.second > y.second;
-              });
-    vocab_size_ = static_cast<int32>(1 + ordered.size());
-    return ordered;
-   }
-
-   std::unordered_map<string, int32>
-   CreateWord2Index(std::vector<WordFreq> vocabulary) {
-    std::unordered_map<string, int32> word_id;
-    Tensor word(DT_STRING, TensorShape({vocab_size_}));
-    Tensor freq(DT_INT32, TensorShape({vocab_size_}));
-    word.flat<string>()(0) = "UNK";
-    int64 total_counted = 0;
-
-    for (std::size_t i = 0; i < vocabulary.size(); ++i) {
-      const auto& w = vocabulary[i].first;
-      auto id = i + 1;
-      word.flat<string>()(id) = w;
-      auto word_count = vocabulary[i].second;
-      freq.flat<int32>()(id) = word_count;
-      total_counted += word_count;
-      word_id[w] = id;
-    }
-
-    freq.flat<int32>()(0) = corpus_size_ - total_counted;
-    vocab_words_ = word;
-    freq_ = freq;
-
-    return word_id;
-   }
-
-   void CreateCorpus(const string& data,
-                     std::unordered_map<string, int32> word_id) {
-
-    static const int32 kUnkId = 0;
-    StringPiece input = data;
-    string w;
-
-    corpus_.reserve(corpus_size_);
-    while (ScanWord(&input, &w)) {
-      corpus_.push_back(gtl::FindWithDefault(word_id, w, kUnkId));
-    }
-
-   }
-
    void CreateCoocurrences() {
+    std::unordered_map<uint64, float> coocurrences_;
     uint64 center_word, context_word, start_index, end_index, dist;
     uint64 index = 0;
     uint64 size = static_cast<uint64>(corpus_.size());
@@ -260,6 +174,7 @@ class GloveModelOp : public OpKernel {
       indices.matrix<int64>()(i, 0) = center_word;
       indices.matrix<int64>()(i, 1) = context_word;
       values.flat<float>()(i) = coocurrences_[index];
+      valid_values.push_back(coocurrences_[index]);
     }
 
     indices_ = indices;
@@ -268,20 +183,7 @@ class GloveModelOp : public OpKernel {
    }
 
    Status Init(Env *env, const string& filename) {
-    string data;
-    TF_RETURN_IF_ERROR(ReadFileToString(env, filename, &data));
-    auto word_freq = CreateWordFrequencies(data);
-
-   // if (corpus_size_ < window_size_ * 10) {
-   //   return errors::InvalidArgument("The text file ", filename,
-   //                                  " contains too little data: ",
-   //                                  corpus_size_, "words");
-   // }
-
-    auto ordered = CreateVocabulary(word_freq);
-    word_freq.clear();
-    auto word_id = CreateWord2Index(ordered);
-    CreateCorpus(data, word_id);
+    WordEmbeddingModel::Init(env, filename);
     CreateCoocurrences();
     precalc_examples_.resize(num_precalc_examples);
